API

Base.Colon

TensorOperator.BaseTensor

TensorOperator.BaseTensor

TensorOperator.Tensor

TensorOperator.Tensor

Base.:*

Base.:+

Base.:-

Base.sum

TensorOperator.:×

TensorOperator.:⊗

TensorOperator.:⋅

TensorOperator.project

TensorOperator.tr

Base.Colon — Method

(::Colon)(e₁::BaseTensor{dim},e₂::BaseTensor{dim}) where dim
(::Colon)(t::Tensor{dim},e::BaseTensor{dim}) where dim
(::Colon)(e::BaseTensor{dim},t::Tensor{dim}) where dim
(::Colon)(t₁::Tensor{dim,order_1,N_1},t₂::Tensor{dim,order_2,N_2}) where {dim,order_1,order_2,N_1,N_2}

Double-dot product for tensors and base tensors. These operaotrs only suit for the tensors that their order greater than or equal to 2. The operations between second-order tensors or base tensors will return a scale value. Other operations will return a low order tensor or base tensor.

TensorOperator.BaseTensor — Type

BaseTensor{dim,order,T}

Type for base tensors, where dim is the dimension, order is the tensor's order, T is the type for its component.

TensorOperator.BaseTensor — Method

BaseTensor(e...)

The constructor for first-order base tensors with components e. The corresponding component is defaultly set to be 1.0. Only first-order base tensors can be implemented by this function, i.e. order = 1, the high order ones can be constructed by dyadic operator ⊗. The dimension of base tensor depends on the length of inputed components, i.e. dim = length(e).

TensorOperator.Tensor — Type

Tensor{dim,order,N,T}

Type for tensor, where dim is the dimension, order is tensor order, N is the number of base tensors in this tensor, T is the data type for component.

TensorOperator.Tensor — Method

Tensor(e::BaseTensor...)

Constructor for tensor. Return a tenor with base tensors of e's.

Base.:* — Method

*(c,e::BaseTensor)

Times operator for a constant and a base tensor. Return a BaseTensor with a component equal to c times the origin component.

Base.:+ — Method

+(e::BaseTensor...)
+(t::Tensor{dim,order},e::BaseTensor{dim,order}) where {dim,order}

Plus operator for tensors and base tensors, where the base tensors should have the same dimension dim, order order and component type T. Return a Tensor onstruced by base tensors e's.

Base.:- — Method

-(t,e::BaseTensor)

Return t plus e that its component is negative.

Base.sum — Method

sum(e::BaseTensor...)

Return a tensor with base tensors e.

TensorOperator.:× — Method

×(v₁::Tuple{Vararg{T,3}},v₂::Tuple{Vararg{T,3}}) where T
×(e₁::BaseTensor{3},e₂::BaseTensor{3})
×(t::Tensor{dim},e::BaseTensor{dim}) where dim
×(e::BaseTensor{dim},t::Tensor{dim}) where dim
×(t₁::Tensor{dim,order_1,N_1},t₂::Tensor{dim,order_2,N_2}) where {dim,order_1,order_2,N_1,N_2}

Cross product for tensors and base tensors. These operaotrs only suit for the tensors that their order greater than or equal to 2. The operations between second-order tensors or base tensors will return a scale value. Other operations will return a low order tensor or base tensor.

TensorOperator.:⊗ — Method

⊗(e₁::BaseTensor{dim},e₂::BaseTensor{dim}) where dim
⊗(t::Tensor{dim},e::BaseTensor{dim}) where dim
⊗(e::BaseTensor{dim},t::Tensor{dim}) where dim
⊗(t₁::Tensor{dim,order_1,N_1},t₂::Tensor{dim,order_2,N_2}) where {dim,order_1,order_2,N_1,N_2}

Dyadic operaotrs for tensors and base tensors. Inputing two base tensors returns a high order base tensor. Inputing at least one tensor returns a tensor with high order base tensors.

TensorOperator.:⋅ — Method

⋅(e₁::BaseTensor{dim},e₂::BaseTensor{dim}) where dim
⋅(t::Tensor{dim},e::BaseTensor{dim}) where dim
⋅(e::BaseTensor{dim},t::Tensor{dim}) where dim
⋅(t₁::Tensor{dim,order_1,N_1},t₂::Tensor{dim,order_2,N_2}) where {dim,order_1,order_2,N_1,N_2}

Dot product for tensors and base tensors. The operations carried between first-order tensors or base tensors will return a scale value. Other operations will return a low order tensor or base tensor.

TensorOperator.project — Method

project(t::Tensor{dim},bases_tensor::Tuple{Vararg{BaseTensor{dim,1},dim}}=(e₁,e₂,e₃)) where dim

Project t to a tensor with the base tensor bases and return it. This function can also be used to simplfy t's base tensors to be independent.

TensorOperator.tr — Method

tr(t::Tensor{3,2})

Return the trace of tensor t. This function is only applicatable for the tensor with dim = 3 and order = 2.